MeetingStationPicker component created, station inputs moved to this component from the form page
