+++ web/src/pages/NewMeeting.jsx
@@ -6,6 +6,8 @@ import { useEffect, useState } from "react";
 import "./NewMeeting.css";
 import { useNavigate } from "react-router-dom";
 
+import MeetingStationPicker from "../components/meetingStationPicker/MeetingStationPicker";
+
 function NewMeeting() {
 	const [stations, setStations] = useState([]);
 	const [helpIconToggle, setHelpIconToggle] = useState(false);
@@ -126,19 +128,19 @@ function NewMeeting() {
 			updatedAttendees[index][field] = value;
 			setAttendees(updatedAttendees);
 		} else if (field === "station") {
-			// ======= Validation: Check if the station matches any meeting station =======
-			const isMeetingStation = copyOfMeetingStations.some(
-				(meetingStation) =>
-					meetingStation.station.station_name.toLowerCase() ===
-					value.toLowerCase(),
-			);
+			// // ======= Validation: Check if the station matches any meeting station =======
+			// const isMeetingStation = copyOfMeetingStations.some(
+			// 	(meetingStation) =>
+			// 		meetingStation.station.station_name.toLowerCase() ===
+			// 		value.toLowerCase(),
+			// );
 
-			if (isMeetingStation) {
-				alert("An attendee's station cannot be the same as a meeting station.");
-				return; // Stop further processing
-			}
-			// ======= End of added validation =======
-			// Update the input value
+			// if (isMeetingStation) {
+			// 	alert("An attendee's station cannot be the same as a meeting station.");
+			// 	return; // Stop further processing
+			// }
+			// // ======= End of added validation =======
+			// // Update the input value
 			const updatedInputValues = [...attendeeInputValues];
 			updatedInputValues[index] = value;
 			setAttendeeInputValues(updatedInputValues);
@@ -194,14 +196,6 @@ function NewMeeting() {
 		setHelpIconToggle(!helpIconToggle);
 	};
 
-	const deleteStation = (e, index) => {
-		e.preventDefault();
-		const updatedMeetingStations = copyOfMeetingStations.filter(
-			(_, i) => i !== index,
-		);
-		setCopyOfMeetingStations(updatedMeetingStations);
-	};
-
 	return (
 		<>
 			<div id="page-container">
@@ -210,143 +204,11 @@ function NewMeeting() {
 					<form onSubmit={handleSubmit}>
 						<h2 className="form-header">Plan your meeting details</h2>
 
-						<div className="form-group">
-							<label id="list-heading" htmlFor="stn-list">
-								Meeting Station List
-							</label>
-							<div id="input-box-container" style={{ marginBottom: "1rem" }}>
-								<ul id="stn-list">
-									{copyOfMeetingStations.map((stationObject, index) => (
-										<li className="meeting-list-item" key={index}>
-											<div className="form-group station-input-group">
-												<input
-													type="text"
-													id={`meeting-station-input-${index}`}
-													value={stationObject.station.station_name || ""}
-													onChange={(e) => {
-														const value = e.target.value;
-														const updatedStations = [...copyOfMeetingStations];
-														const exactMatch = stations.find(
-															(station) => station.station_name === value,
-														);
-
-														if (exactMatch) {
-															updatedStations[index].station = exactMatch;
-														} else {
-															updatedStations[index].station = {
-																station_name: value,
-															};
-														}
-														setCopyOfMeetingStations(updatedStations);
-													}}
-													list={`meeting-stations-list-${index}`}
-													placeholder=" "
-													required
-												/>
-												<datalist id={`meeting-stations-list-${index}`}>
-													{stations
-														.filter((station) =>
-															station.station_name
-																.toLowerCase()
-																.includes(
-																	(
-																		stationObject.station.station_name || ""
-																	).toLowerCase(),
-																),
-														)
-														.map((station, idx) => (
-															<option key={idx} value={station.station_name}>
-																{station.crs_code}
-															</option>
-														))}
-												</datalist>
-												<label htmlFor={`meeting-station-input-${index}`}>
-													Meeting Station {index + 1}
-												</label>
-											</div>
-											<button
-												onClick={(e) => deleteStation(e, index)}
-												className="delete-button"
-												aria-label={`Delete station ${index + 1}`}
-												type="button"
-											>
-												<svg
-													xmlns="http://www.w3.org/2000/svg"
-													width="30"
-													height="30"
-													viewBox="0 0 64 64"
-												>
-													<rect
-														x="16"
-														y="22"
-														width="32"
-														height="34"
-														fill="red"
-														rx="4"
-													/>
-													<rect
-														x="12"
-														y="16"
-														width="40"
-														height="5"
-														fill="black"
-														rx="2"
-													/>
-													<rect
-														x="24"
-														y="11"
-														width="16"
-														height="6"
-														fill="inherit"
-														rx="2"
-													/>
-													<rect
-														x="22"
-														y="27"
-														width="4"
-														height="24"
-														fill="white"
-													/>
-													<rect
-														x="30"
-														y="27"
-														width="4"
-														height="24"
-														fill="white"
-													/>
-													<rect
-														x="38"
-														y="27"
-														width="4"
-														height="24"
-														fill="white"
-													/>
-												</svg>
-											</button>
-										</li>
-									))}
-								</ul>
-
-								<button
-									id="add-station-button"
-									type="button"
-									onClick={(e) => {
-										e.preventDefault();
-										setCopyOfMeetingStations([
-											...copyOfMeetingStations,
-											{ station: { station_name: "" } },
-										]);
-									}}
-									aria-label="Add station"
-									name="add-station"
-								>
-									<span style={{ fontSize: "18px", marginRight: "5px" }}>
-										+
-									</span>{" "}
-									Add Station
-								</button>
-							</div>
-						</div>
+						<MeetingStationPicker
+							stations={stations}
+							copyOfMeetingStations={copyOfMeetingStations}
+							setCopyOfMeetingStations={setCopyOfMeetingStations}
+						/>
 
 						<div className="form-group">
 							<input
